Fixed add to equipment services, stats calculated after add item to equipment, fixed item lvl to type double and liquibase scripts

+++ src/main/kotlin/com/game/dungeonborn/dto/item/ItemDTO.kt
@@ -9,7 +9,7 @@ data class ItemDTO(
     var name: String,
     var type: ItemType,
     var slot: SlotType,
-    var itemLevel: Int,
+    var itemLevel: Double,
     var quality: ItemQuality,
     var stamina: Double,
     var strength: Double,
